tipado fix2

+++ common/core/memory_subsystem/page_stats.h
@@ -11,8 +11,8 @@ public:
     void updatePageStats(IntPtr pageAdress, bool evict = false);
 
 private:
-    std::unordered_map<int, std::bitset<1024>> pageMap; // Asumimos que el tamano de la pagina es 4096 bytes (1024 palabras de 4 bytes)
-    std::unordered_map<int, int> pageAccessCount;
+    std::unordered_map<IntPtr, std::bitset<1024>> pageMap; // Asumimos que el tamano de la pagina es 4096 bytes (1024 palabras de 4 bytes)
+    std::unordered_map<IntPtr, int> pageAccessCount;
     UInt64 totalPages;
     UInt64 minUsage;
     UInt64 maxUsage;
